chore: address pull request feedback
