+++ packages/api/src/router/widgets/health-monitoring.ts
@@ -1,6 +1,6 @@
 import { observable } from "@trpc/server/observable";
 
-import type { HealthMonitoring } from "@homarr/integrations";
+import type { SystemHealthMonitoring } from "@homarr/integrations";
 import type { ProxmoxClusterInfo } from "@homarr/integrations/types";
 import { clusterInfoRequestHandler, systemInfoRequestHandler } from "@homarr/request-handler/health-monitoring";
 
@@ -28,7 +28,7 @@ export const healthMonitoringRouter = createTRPCRouter({
   subscribeSystemHealthStatus: publicProcedure
     .concat(createManyIntegrationMiddleware("query", "openmediavault", "dashDot", "mock"))
     .subscription(({ ctx }) => {
-      return observable<{ integrationId: string; healthInfo: HealthMonitoring; timestamp: Date }>((emit) => {
+      return observable<{ integrationId: string; healthInfo: SystemHealthMonitoring; timestamp: Date }>((emit) => {
         const unsubscribes: (() => void)[] = [];
         for (const integration of ctx.integrations) {
           const innerHandler = systemInfoRequestHandler.handler(integration, {});
