+++ packages/integrations/src/index.ts
@@ -27,14 +27,17 @@ export type { IntegrationInput } from "./base/integration";
 export type { DownloadClientJobsAndStatus } from "./interfaces/downloads/download-client-data";
 export type { ExtendedDownloadClientItem } from "./interfaces/downloads/download-client-items";
 export type { ExtendedClientStatus } from "./interfaces/downloads/download-client-status";
-export type { SystemHealthMonitoring as HealthMonitoring } from "./interfaces/health-monitoring/health-monitoring-types";
+export type { SystemHealthMonitoring } from "./interfaces/health-monitoring/health-monitoring-types";
 export { MediaRequestStatus } from "./interfaces/media-requests/media-request-types";
 export type { MediaRequestList, MediaRequestStats } from "./interfaces/media-requests/media-request-types";
 export type { StreamSession } from "./interfaces/media-server/media-server-types";
-export type { TdarrQueue } from "./interfaces/media-transcoding/queue";
-export type { TdarrPieSegment, TdarrStatistics } from "./interfaces/media-transcoding/statistics";
-export type { TdarrWorker } from "./interfaces/media-transcoding/workers";
-export type { Notification } from "./interfaces/notifications/notification";
+export type {
+  TdarrQueue,
+  TdarrPieSegment,
+  TdarrStatistics,
+  TdarrWorker,
+} from "./interfaces/media-transcoding/media-transcoding-types";
+export type { Notification } from "./interfaces/notifications/notification-types";
 
 // Schemas
 export { downloadClientItemSchema } from "./interfaces/downloads/download-client-items";
