+++ packages/integrations/src/interfaces/media-transcoding/media-transcoding-types.ts
@@ -1,3 +1,20 @@
+export interface TdarrQueue {
+  array: {
+    id: string;
+    healthCheck: string;
+    transcode: string;
+    filePath: string;
+    fileSize: number;
+    container: string;
+    codec: string;
+    resolution: string;
+    type: "transcode" | "health-check";
+  }[];
+  totalCount: number;
+  startIndex: number;
+  endIndex: number;
+}
+
 export interface TdarrPieSegment {
   name: string;
   value: number;
@@ -20,4 +37,18 @@ export interface TdarrStatistics {
   videoResolutions: TdarrPieSegment[];
   audioCodecs: TdarrPieSegment[];
   audioContainers: TdarrPieSegment[];
+}
+
+export interface TdarrWorker {
+  id: string;
+  filePath: string;
+  fps: number;
+  percentage: number;
+  ETA: string;
+  jobType: string;
+  status: string;
+  step: string;
+  originalSize: number;
+  estimatedSize: number | null;
+  outputSize: number | null;
 }
