--- packages/integrations/src/interfaces/media-transcoding/queue.ts
@@ -1,16 +1,0 @@
-export interface TdarrQueue {
-  array: {
-    id: string;
-    healthCheck: string;
-    transcode: string;
-    filePath: string;
-    fileSize: number;
-    container: string;
-    codec: string;
-    resolution: string;
-    type: "transcode" | "health-check";
-  }[];
-  totalCount: number;
-  startIndex: number;
-  endIndex: number;
-}
