+++ packages/integrations/src/interfaces/notifications/notification-types.ts
