+++ packages/integrations/src/interfaces/notifications/notifications-integration.ts
@@ -1,4 +1,4 @@
-import type { Notification } from "./notification";
+import type { Notification } from "./notification-types";
 
 export interface INotificationsIntegration {
   getNotificationsAsync(): Promise<Notification[]>;
