+++ packages/integrations/src/ntfy/ntfy-integration.ts
@@ -4,7 +4,7 @@ import { ResponseError } from "@homarr/common/server";
 import { Integration } from "../base/integration";
 import type { IntegrationTestingInput } from "../base/integration";
 import type { TestingResult } from "../base/test-connection/test-connection-service";
-import type { Notification } from "../interfaces/notifications/notification";
+import type { Notification } from "../interfaces/notifications/notification-types";
 import type { INotificationsIntegration } from "../interfaces/notifications/notifications-integration";
 import { ntfyNotificationSchema } from "./ntfy-schema";
